coins

+++ lib/store/dogmodel.dart
@@ -2,7 +2,7 @@ class Product {
   int id;
   String title;
   String imgUrl;
-  double price;
+  int price;
   int qty;
 
   Product({this.id, this.title, this.price, this.qty, this.imgUrl});
